[fix] don't append span on metrics endpoint
