+++ telemetry.go
@@ -1,7 +1,11 @@
 package drudge
 
 import (
+	"fmt"
+	"net/http"
+
 	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 	"github.com/pkg/errors"
 	"github.com/uber/jaeger-client-go"
 	jaegercfg "github.com/uber/jaeger-client-go/config"
@@ -72,4 +76,30 @@ func Jaeger(c interface{}) (func(), error) {
 	return func() {
 		_ = closer.Close()
 	}, nil
+}
+
+var drudgeTag = opentracing.Tag{Key: string(ext.Component), Value: "drudge"}
+
+func tracingWrapper(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/metrics" {
+			h.ServeHTTP(w, r)
+			return
+		}
+
+		parentSpanContext, err := opentracing.GlobalTracer().Extract(
+			opentracing.HTTPHeaders,
+			opentracing.HTTPHeadersCarrier(r.Header),
+		)
+		if err == nil || err == opentracing.ErrSpanContextNotFound {
+			serverSpan := opentracing.GlobalTracer().StartSpan(
+				fmt.Sprintf("http.%s.[%s]", r.Method, r.URL.Path),
+				ext.RPCServerOption(parentSpanContext),
+				drudgeTag,
+			)
+			r = r.WithContext(opentracing.ContextWithSpan(r.Context(), serverSpan))
+			defer serverSpan.Finish()
+		}
+		h.ServeHTTP(w, r)
+	})
 }
